refactor(workstation): extract cursor-sync + load-more pagination into hooks (0.72 app.ts decomposition)
